[CLOB-1036] organizational update for liquidation daemon
