+++ protocol/daemons/liquidation/client/grpc_helper_test.go
@@ -3,6 +3,8 @@ package client_test
 import (
 	"context"
 	"errors"
+	"testing"
+
 	"github.com/cometbft/cometbft/libs/log"
 	"github.com/cosmos/cosmos-sdk/types/query"
 	"github.com/dydxprotocol/v4-chain/protocol/daemons/flags"
@@ -14,7 +16,6 @@ import (
 	clobtypes "github.com/dydxprotocol/v4-chain/protocol/x/clob/types"
 	satypes "github.com/dydxprotocol/v4-chain/protocol/x/subaccounts/types"
 	"github.com/stretchr/testify/require"
-	"testing"
 )
 
 func TestGetAllSubaccounts(t *testing.T) {
